Move CryptoMarketPricesReceiver and CryptoMarketPricesReceiverDelegate to a file
